requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
python-dotenv==1.0.1
aiofiles==23.2.1
celery==5.4.0
redis==5.0.6
Pillow==10.3.0
requests==2.32.3
pydantic==2.7.1
pydantic-settings==2.2.1
ffmpeg-python==0.2.0
boto3==1.34.111
google-cloud-texttospeech==2.16.3
elevenlabs==1.2.0
stability-sdk==0.8.5
